Added a lot of assert(cur_bp) lines.
Reuben has spontaneously decided that Zee should be able to cope with
the case in which there are no buffers. For various reasons it is now
quite easy to construct such a case, but large chunks of the program
can't yet cope. In other words, we suddenly have a large number of
bugs that we previously didn't. The number is not far short of 100.

On Reuben's suggestion, I have guarded lots of parts of the program with
assertions. This does not fix any of the bugs, but it improves their
symptoms and makes them trivial to find. I've been fairly simple-minded
about this process; some of the assertions are certainly unnecessary but
I felt that false positives were errors on the right side.

The job is not finished yet. So far I've only defended against
cur_bp==NULL. It is also necessary to defend against cur_wp==NULL at
least.

	Alistair