Make some minor improvements to key handling:
Make keytostr do linear search, not binary. There are so few named
keys this is not a problem, and has the advantage that key codes no
longer need to be in alphabetical order, which is easier to maintain.

Add KBD_SPC code for space, which has its own symbol in bindings, SPC.

Redo key arrays (keyname and keycode) in same order as keys in main.h;
in future, combine these into tbl_keys.h that can be reused just just
the other tbl_*.h headers.