FIXME attack.
Use vector for bindings. Don't bother keeping them in sorted order, as
we're using a linear search anyway.

Make where-is find all the bindings, not just the first.

Make compact_path take an astr, not a char *.

Make resync_display assert cur_wp. It's OK to assert cur_bp here for
the moment.

Make insert_nstring take an astr.

Add file_read to common up file_open and file_insert.

Determine the EOL string of the file as it's read in. Note that
file_insert still hasn't been changed to work properly.

Make raw_write_to_disk use only ANSI functions.

Rename copy-region-as-kill to copy-region.

Make line_replace_text return whether anything was changed rather than
twiddling the buffer's flags; this means it doesn't need cur_bp.

Add a call to lisp_parse_end in lisp_parse.

Always print out the result of any Lisp interpreted because of
command-line flags -l or -e to stdout, so that it goes somewhere when
we're not running in batch mode.

Fix the memory leak in FUNCALL by not passing a uniarg (we don't need
to, duh). The leak in FUNCALL_ARG remains.

In the Lisp parser, tabs now round up to the next multiple of 8
spaces.

Fix a leak in the Lisp parser.

#include string.h in vector.c for memmove.