Add vec_shrink to remove part of a vector, and vec_offset to return an index given an object in a vector.
Make the static function resize return void; it didn't need to return
anything, as it always returned its argument; change the one place
where the return value was used.

Reformat the #defines in vector.h to be more consistent with the rest
of Zee.

Extend the copyright backwards in time as well as forwards, to more
correctly reflect reality. Since I've deleted the original nuweb
sources, I don't know precisely when the code was started, grr...