Fix kill_line to use delete_nstring rather than repeated calls to delete-char.
Make delete_nstring append EOLs to the returned string according to
the buffer's EOL string, and make the calls that reinsert text use the
same string. This doesn't change anything at the moment, but if we use
the kill buffer for something else later on it will mean the string cut
is accurate.

Comment delete_nstring.