Fold create_buffer into buffer_new.
Don't worry about filenames any more, since we're only editing one
file at a time. Remove file-insert (a special case of shell-command)
and file-save-as (this is really an undo problem, not an editor
problem). As a result, get rid of all the machinery for dealing with
filenames and paths, and get rid of a lot of dependencies on non-ANSI
functions.

Simplify draw_status_line.

Fold open_file_at into main.

Assert calls to calculate_the_region, and rename calculate_region to
region_size and make it static.

Get rid of term_full_redisplay, which was only called once and was
trivial.