Add root_varlist for the case where there is no buffer. Add init_variables and free_variables to initialise and destroy it.
Tidy up main logic, so we don't exit early if we print the help
message, batch mode no longer forces the user init file not to be
loaded, and the terminal is only written to if there is a buffer and
hence a window. Also cure some bugs related to +LINE FILE notation.

Remove lingering references to Lisp.