Start moving towards proper command wrapper declarations.
We now have

  DEF_ARG(cmd, doc, arg1, ...)

to declare commands that take arguments.

  DEF_INT(cmd, doc)

becomes

  DEF_ARG(cmd, doc, INT(intarg))

This will be expanded to other types of argument. In the end, all the
command definitions will be in a (hand written) tbl_funcs.h, and
although the C code will still have to use a special calling
convention (to allow functions to be called anonymously independent of
type), the code to obtain arguments (whether interactively or from the
argument list) won't have to be repeated in each function.