Gradually moving minibuf.c in the direction of comprehensibility. One or two simplifications consequent on Reuben's earlier rationalisation of the COMPLETION_XXX return types. A switch became an if, for example.
I'm having difficulty understanding the state space of
minibuf_read_completion. There are definitely bugs in e.g. whether the
popup is shown or not. It is not clear how to test if the popup is shown
or not. Some bits of code use "lasttab" while others use
COMPLETION_POPPED_UP. Needs rationalising.