Fixed rblist_split. I replaced the while loop with a recursive function. Pity, but maybe we can optimise it later.
Discovered and fixed a bug in rblist_iterate. Didn't work for empty the list.

Added tests to find the above bugs in future.

rblist.c is not close to specification. Still missing are rblist_compare
and conversion of character positions to/from line numbers.

I'm having second thoughts about generification. I think I'm going to
commit this data structure to the element type "char" for the moment. It
makes several things easier.