Rewrote outch, draw_line and draw_window among others. The horizontal scroll offset is now a column (which makes sense) not a character position (which doesn't). calculate_start_column is still b0rk3d due to not appreciating the difference, and needs further attention.
In the process, we (apt1002 and rrt) removed a lot of things that were
hangovers of the days of multiple windows, such as cur_topline, and we
rationalised things a bit.

We also found a bug whereby the region was not drawn properly if it
finished just after the start of a line. The desired behaviour is that
the whole of the right-hand end of the previous line (all unused space
to the edge of the screen) should be highlighted. The observed behaviour
was that only part of the unused space was highlighted; the size of the
highlighted part was equal to the number of characters highlighted on
the following line. Now fixed.

The result works no worse than it did before, which is nice. :-) The
cursor is still not drawn in the right place, and the horizontal
scrolling is a bit suspicious in the presence of tab characters.