The main change was to rewrite calculate_start_column() as per plan. The detailed specification is given in the comment, but the basic idea is that the result depends only on the width of the screen, the column of the cursor and the column of the end of the line. At no point is the character position of the cursor or the end of the line relevant.
The result now basically works, reassuringly. The remaining bugs are not
display bugs. :-)

Fixed a bug in draw_line() whereby the $ on the right-hand end of the
line did not take account of horizontal scrolling.

Thought better of a FIXME in draw_window().

Commented that Alt-Tab is a useless key binding in most window managers,
but didn't remove it.

Removed the key-binding for the Tab key (indent_relative), so as to be
able to insert tab characters without using edit_insert_quoted. This
makes zee more like e.g. Notepad. Users who don't know about
indent_relative will no longer be confused; those that do know and care
about it will cope.