Made a new "rbutil" file, containing rblist_fmt, make_string_printable and make_char_printable:
 - rblist_fmt comes straight out of rblist.c. Moving it breaks a
dependency cycle.

 - make_string_printable is an amalgum of make_char_printable,
display_string_width and covert_goal_to_column.

 - make_char_printable is a stub that wraps make_string_printable.

The old functions column_to_character and string_display_width are now
reimplemented (with no change of contract) as special cases of
make_string_printable, and are very short. They could probably be
in-lined and removed.

The result of this change is not really much better by most metrics. The
code is now a bit shorter, but not by much, and it could be argued that
the saving was through making individual functions longer. The code
should now be slightly faster, but annoyingly it is still necessary to
call make_char_printable in a tight loop. The main benefit of the change
is that everything related to converting strings and positions to
printable strings and screen columns is in one function
(make_string_printable).

Removed the function outch in term.c. outch_printable remains with two
call sites both in draw_line. It is very short, and could be in-lined
and removed too.

Found a bug in in_region which I don't understand; see FIXME.

Found several bugs in draw_line related to doing things in the wrong
order. For example, if there was a $ at the left-hand end of the line
and the selection included the line ending, the spaces printed at the
end of the line were actually overwriting the beginning of the line
because the cursor position had been changed. For similar a reason they
were in the wrong colour. These bugs are now all fixed.