Rewrite history mechanism in Lua.
Use clue (clue.h) for binding to Lua. Lua bindings are a bit of a mess
at the moment, a mixture of clue (in minibuf.c, for the history
functions) and ad-hoc (elsewhere), but as I keep saying, it will all
go away. For the mean time, I'll try to use clue wherever I make new
bindings.

Make variable functions use ordinary strings rather than rblists for
easier impedance matching with Lua.

Use 0xff, not 255, when talking about character codes that fit in a
byte.