Add my Lua standard library, std.lua (the all-in-one-file mix). It's silly not to avail ourselves of the many useful functions, and it doesn't count as Zee code, oh no. Use stdlib's die in the mk*.lua scripts.
Remove the implementation of lists as Lua tables, which is no longer
needed; all lists are in Lua.

Count LOC in one place only for an honest overall view of the amount
of code written for the project.

Fix FIXME for checking return of CMDCALL_UINT in process_key.

Allow texinfo markup in tbl_vars.lua, translated appropriately for
plain text output.

Make macros out of commands rather than keys.