keycode: make sure to only pass the key string to keycode().
* src/keycode (keycode_mt:__sub): gsub returns a tuple, which
was being passed to the memoized keycode function. Now that
memoize use all arguments as a key, the extra gsub return value
was causing the wrong value to be returned.  This fix takes care
to discard the unused second return value of gsub before calling
keycode so that the correct value can be looked up in the cache
by memoize.