registers: fix tests/interactive/copy-to-register.el.
After changing the internal key representation, we also need a
way to convert from an interactive unfiltered keypress (e.g. in
Emacs \C-xrx\F1 saves to register ESC, and inserts OP into the
current buffer) to a register entry key, while at the same time
still supporting the (copy-to-register) lisp argument being a
byte (e.g. (copy-to-register 49 ...) saves to register '1').
* src/term_curses.lua (term_bytetokey): New function, the
inverse of term_keytobyte.
* src/term/registers.lua (copy-to-register): Use
getkey_unfiltered in interactive mode so that a raw byte is
returned as per Emacs, and then use term_bytetokey to convert
that byte to a register key.
(insert_register, insert-register): Use it to get the register
key.
(write_registers_list): Simplify now the we're using a keycode
as the register list key.