funcs: fix tests/interactive/shell-command.el.
* src/funcs.el (shell-command): When called interactively with a
prefix argument, the lisp function calling discipline passes
that prefix argument directly to the command as it's first
conventional argument.  So \M-1\M-xshell-command\rls\r is
equivalent to (shell-command 1 nil) not (shell-command "ls" t).
Detect this by checking whether the first argument matches the
last prefix arg and the second argument is nil, and adjust
accordingly.