curses: support \C-? read-syntax for compatibility with Emacs.
Fix the last few failing tests, which are using \C-? to invoke
backward-delete-char, which works already with Emacs. \C-? is a
reliable way to pass ASCII 0x7f to execute-kbd-macro, regardless
of what terminfo says about the kbs capability (which becomes
\BACKSPACE).
* src/term_curses (term_init): Add an entry for \C-? read-syntax
to the key translation maps.
* src/default-bindings.el: Bind \C-? and \M-C-? to the same
commands as \BACKSPACE and \M-\BACKSPACE respectively.