minibuf: make sure the cursor is displayed properly.
The recent optimisation for not redrawing the screen during
typeahead catchup has the side-effect of never leaving the cursor
in minibuf, even when typing in the minibuffer!
* src/getkey.lua (getkeystroke): Split out the key fetching part
of the optimised getkey() function.
(getkey): Adjust to call getkeystroke() instead of raw term_getkey()
keeping typeahead optimisation.
* src/minibuf.lua (minibuf_read_yn): Use getkeystroke() to make
sure the cursor is not moved out of minibuf during term_refresh().
* src/term_minibuf.lua (term_minibuf_read): Likewise.