configury: distribute zile.1.in correctly.
* doc/Makefile.in (MAINTAINERCLEANFILES): zile.1.in is a distributed
file, so it will be in $(srcdir) not $(builddir).
(DISTCLEANFILES): Don't clean zile.1.in, because it is distributed!
(doc/zile.1): Force the dependency to be on zile.1.in in $(srcdir).
(doc/zile.1.in): Name $(srcdir) explicitly, for non-gmake.
Also, exit gracefully when it's not possible to regenerate zile.1.in
in a read-only directory, such as at `make dist` time.